+++ src/main/java/com/seoultech/ecc/expression/repository/ExpressionRepository.java
@@ -3,5 +3,5 @@ package com.seoultech.ecc.expression.repository;
 import com.seoultech.ecc.expression.datamodel.ExpressionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpressionRepository extends JpaRepository<ExpressionEntity, Long> {
+public interface ExpressionRepository extends JpaRepository<ExpressionEntity, Integer> {
 }
